requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
azure-core==1.38.2
azure-functions==1.24.0
azure-identity==1.25.2
build==1.4.0
bump2version==1.0.1
cachetools==7.0.1
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.4
cryptography==46.0.5
distlib==0.4.0
dnspython==2.8.0
docutils==0.22.4
email-validator==2.3.0
execnet==2.1.2
fastapi==0.135.1
filelock==3.25.0
greenlet==3.3.2
h11==0.16.0
id==1.6.1
identify==2.6.17
idna==3.11
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
JayDeBeApi==1.2.3
jeepney==0.9.0
jpype1==1.6.0
keyring==25.7.0
librt==0.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msal==1.35.0
msal-extensions==1.3.1
mssql-python==1.4.0
mypy==1.19.1
mypy_extensions==1.1.0
nh3==0.3.3
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.5.1
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyodbc==5.3.0
pyproject-api==1.10.0
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-discovery==1.1.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
readme_renderer==44.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.3
ruff==0.15.4
SecretStorage==3.5.0
SQLAlchemy==2.0.47
sqlalchemy-jdbcapi==2.2.1
sqlmodel==0.0.37
starlette==0.52.1
tox==4.47.0
twine==6.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==21.1.0
Werkzeug==3.1.6
